fixing include paths for Arduino IDE
