+++ software/src/apps/DQ.h
@@ -26,17 +26,7 @@
 // from Braids by Olivier Gillet (see braids_quantizer.h/cc et al.). It has since
 // grown a little bit...
 
-#include "OC_apps.h"
-#include "util/util_settings.h"
-#include "util/util_trigger_delay.h"
-#include "OC_menus.h"
-#include "OC_visualfx.h"
-#include "OC_scales.h"
-#include "OC_scale_edit.h"
-#include "OC_strings.h"
-#include "OC_digital_inputs.h"
-#include "OC_ADC.h"
-#include "src/extern/dspinst.h"
+#pragma once
 
 extern uint_fast8_t MENU_REDRAW;
