+++ software/src/apps/TheDarkestTimeline.h
@@ -22,24 +22,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
-#include <Arduino.h>
-#include <EEPROM.h>
-#include <stdint.h>
-#include "OC_core.h"
-#include "OC_config.h"
-#include "OC_apps.h"
-#include "OC_patterns.h"
-#include "OC_strings.h"
-#include "OC_menus.h"
-#include "OC_ui.h"
-#include "src/drivers/display.h"
-
-#include "HemisphereApplet.h"
-#include "util/util_settings.h"
-#include "OC_DAC.h"
-#include "OC_scales.h"
-#include "HSApplication.h"
-#include "HSMIDI.h"
+#pragma once
 
 #define DT_CV_TIMELINE 0
 #define DT_PROBABILITY_TIMELINE 1
