+++ software/src/src/Audio/SafeNoteFrequencyAnalyzer.h
