+++ software/src/src/Audio/synth_advanced_karplus.cpp
